Add VPA memory and cpu limits to seed and shoot Prometheus instances (#8694)
In the past, we've encountered situations where Prometheus experienced
sudden spikes in memory usage. This led the Vertical Pod Autoscaler (VPA)
or Horizontal VPA (HVPA) to set the memory request for Prometheus at a
level exceeding the node's limits, rendering the pod unschedulable. To
avoid this issue, this commit adds the `maxAllowed` field to the (h)vpas
with a value lower than the upper limit that Prometheus should reach during
normal operations, yet equal to or less than the available memory on
smaller nodes.